refactored quiz into reusable component for sub quizes
